Branch for develop StageScene
