+++ Application/src/main/java/io/github/antijava/marjio/scene/sceneObject/IInterruptable.java
@@ -1,0 +1,10 @@
+package io.github.antijava.marjio.scene.sceneObject;
+
+import io.github.antijava.marjio.common.IInput;
+
+/**
+ * Created by firejox on 2015/12/29.
+ */
+public interface IInterruptable {
+    boolean check(IInput input);
+}
