+++ Application/src/main/java/io/github/antijava/marjio/scene/sceneObject/Item.java
@@ -1,0 +1,7 @@
+package io.github.antijava.marjio.scene.sceneObject;
+
+/**
+ * Created by firejox on 2015/12/28.
+ */
+public class Item {
+}
